requirements.txt
requests
beautifulsoup4
tqdm
selenium

python-dotenv
openai